+++ data-features/feature-api/src/commonMain/kotlin/com/grippo/data.features.api/training/models/Iteration.kt
@@ -2,6 +2,6 @@ package com.grippo.data.features.api.training.models
 
 public data class Iteration(
     val id: String?,
-    val weight: Double,
+    val weight: Float,
     val repetitions: Int
 )
