+++ data-services/network/src/commonMain/kotlin/com/grippo/network/dto/ExerciseVolumeDto.kt
@@ -12,5 +12,5 @@ public data class ExerciseVolumeDto(
     @SerialName("id")
     val id: String? = null,
     @SerialName("volume")
-    val volume: Double? = null
+    val volume: Float? = null
 )
